Port the 'my profile' section to React

+++ app/graphql/mutations/update_user_con_profile.rb
@@ -7,9 +7,11 @@ class Mutations::UpdateUserConProfile < Mutations::BaseMutation
   def resolve(**args)
     user_con_profile = convention.user_con_profiles.find(args[:id])
     user_con_profile_attrs = args[:user_con_profile].to_h.stringify_keys
-    user_con_profile.assign_form_response_attributes(
-      JSON.parse(user_con_profile_attrs.delete('form_response_attrs_json'))
-    )
+    if user_con_profile_attrs.key?('form_response_attrs_json')
+      user_con_profile.assign_form_response_attributes(
+        JSON.parse(user_con_profile_attrs.delete('form_response_attrs_json'))
+      )
+    end
     user_con_profile.assign_attributes(user_con_profile_attrs)
     user_con_profile.save!
